tools/requirements.txt
anthropic>=0.96.0
edge-tts>=7.0.0
opencc-python-reimplemented>=0.1.7
